move_particles_diffuse improved
move_particles_diffuse function now works. One last issue with it:
- it seems that the array wraps around both horizontally and vertically right now, meaning that nutrients can go from left to right but also from top to bottom (not very realistic)

The init-array function is meant to generate an array where the nutrients are placed randomly on a grid. Right now the function does not place any nutrients on the top row, which is quite weird. Esther mentioned we should look at this after discussing how we want to initialize the 'stream of nutrients'. Let's look at that one tomorrow.